bugfix: method_subtract error created nested jsonrpc_error_object
it is now the method's responsibility to create its own
jsonrpc_error_object on error. the only valid return values are 0 and -1.

there is no need for jsonrpc_error_object() function to take in a
user-supplied message. any user message could be supplied in the "data"
parameter. See method_subtract for an example.